+++ 2018/src/Fri3dBadge/src/Matrix.h
@@ -9,8 +9,8 @@
 class Matrix {
   public:
     Matrix();
-    void set(int i, int y);
-    void clear(int i, int y);
+    void set(uint8_t i, uint8_t y);
+    void clear(uint8_t i, uint8_t y);
     void clear();
 
     void render(); // called from refresh_screen
@@ -22,14 +22,14 @@ class Matrix {
       }
     }
     void start();
-    void write(char c, int x);
+    void write(char c, uint8_t x);
     void write(char c);
 
 	private:
     // screen is encoded as a byte per column
-    int matrix[14] = { 0 };
+    uint8_t matrix[14] = { 0 };
     void render(int right, int left);
-    int pixel(int x, int y, int eye);
+    int pixel(uint8_t x, uint8_t y, uint8_t eye);
 };
 
 #endif
